### 2018-12-01 #1113 @whedon
[PRE REVIEW]: PsychroLib: a library of psychrometric functions to calculate thermodynamic properties of air
**Submitting author:** @dmey (<a href="http://orcid.org/0000-0002-7071-7547">D. Meyer</a>)
**Repository:** <a href="https://github.com/psychrometrics/psychrolib" target ="_blank">https://github.com/psychrometrics/psychrolib</a>
**Version:** 2.0.0
