### 2018-12-01 #1113 @kyleniemeyer
Hi @dmey, could you try submitting with a non-SVG figure? Perhaps PDF or PNG?
