+++ paper/paper.md
@@ -32,7 +32,7 @@ The estimation of psychrometric properties of air is critical in several enginee
 
 Here, we present PsychroLib, a common set of psychrometric software libraries, aimed at improving scientific reproducibility, reducing the likelihood of software errors, and saving time to scientists and engineers when developing software and working with psychrometric calculations. PsychroLib is a free and open-source psychrometric library, currently based on formulae from the ASHRAE Handbook Fundamentals [@Ashrae2017_IP; @Ashrae2017_SI] for both imperial (IP) and metric (SI) systems of units. It includes common functions for estimating dry, moist, saturated properties of air, and standard atmosphere, such as converting between dry-, wet-, dew-point temperature, relative humidity, humidity ratio and vapour pressure (Figure 1).
 
-![Relationships of common functions as implemented in PsychroLib. Bold arrows show the relationship between function involving a direct call while light arrow show the relationship between two or more. For a complete list of functions available in PsychroLib, see the README file in the project’s repository.](../assets/psychrolib-relationships.svg)
+![Relationships of common functions as implemented in PsychroLib. Bold arrows show the relationship between function involving a direct call while light arrow show the relationship between two or more. For a complete list of functions available in PsychroLib, see the README file in the project’s repository.](psychrolib-relationships.pdf)
 
 PsychroLib is available for Python, C, Fortran, JavaScript, and Microsoft Excel Visual Basic for Applications (VBA). It is developed with a common application programming interface (API) across all the supported languages. All functions have been unit tested with a combination of manual and automated tests against standard ASHRAE reference tables or third-party implementations. PsychroLib is available on GitHub at https://github.com/psychrometrics/psychrolib and released under the MIT licence. We strongly encourage users to provide feedback, bug reports and feature requests, through the GitHub’s issue system at https://github.com/psychrometrics/psychrolib/issues.
 
