### 2018-12-01 a94c723
add padding to paper fig + fix paper refs
