+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-The estimation of psychrometric properties of air is critical in several engineering and scientific applications such as heating, ventilation, and air conditioning (HVAC) and meteorology. Although formulae to calculate the psychrometric properties of air are widely available in the literature (e.g. [@Stull2011; @Wexler1983; @Stoecker1982; @Dilley1968; @Humphreys1920]), their implementation in computer programs or spreadsheets can be challenging and time consuming. To our knowledge, only few numerical implementations of such formulae are freely available as standalone software libraries for programming languages and spreadsheets used in science and engineering.
+The estimation of psychrometric properties of air is critical in several engineering and scientific applications such as heating, ventilation, and air conditioning (HVAC) and meteorology. Although formulae to calculate the psychrometric properties of air are widely available in the literature [@Stull2011; @Wexler1983; @Stoecker1982; @Dilley1968; @Humphreys1920], their implementation in computer programs or spreadsheets can be challenging and time consuming. To our knowledge, only few numerical implementations of such formulae are freely available as standalone software libraries for programming languages and spreadsheets used in science and engineering.
 
 Here, we present PsychroLib, a common set of psychrometric software libraries, aimed at improving scientific reproducibility, reducing the likelihood of software errors, and saving time to scientists and engineers when developing software and working with psychrometric calculations. PsychroLib is a free and open-source psychrometric library, currently based on formulae from the ASHRAE Handbook Fundamentals [@Ashrae2017_IP; @Ashrae2017_SI] for both imperial (IP) and metric (SI) systems of units. It includes common functions for estimating dry, moist, saturated properties of air, and standard atmosphere, such as converting between dry-, wet-, dew-point temperature, relative humidity, humidity ratio and vapour pressure (Figure 1).
 
