### 2018-12-01 #1113 @dmey
Add padding to paper fig + fix paper refs
