### 2018-12-01 1bf449d
move paper fig
