### 2018-12-01 ee44ef4
fix corporate author for bibtex
