+++ paper/paper.bib
@@ -52,7 +52,7 @@
 
 @book{Ashrae2017_IP,
   title={ASHRAE Handbook Fundamentals 2017: Inch-Pound System Edition},
-  author={American Society of Heating Refrigeration and Air Conditioning Engineers},
+  author={{American Society of Heating Refrigeration and Air Conditioning Engineers}},
   year={2017},
   series={ASHRAE Handbook Fundamentals Inch-Pound System},
   publisher={ASHRAE},
@@ -62,7 +62,7 @@
 
 @book{Ashrae2017_SI,
   title={ASHRAE Handbook Fundamentals 2017: International System Edition},
-  author={American Society of Heating Refrigeration and Air Conditioning Engineers},
+  author={{American Society of Heating Refrigeration and Air Conditioning Engineers}},
   year={2017},
   series={ASHRAE Handbook Fundamentals International System},
   publisher={ASHRAE},
