### 2018-12-01 4174e1d
fix corporate author for bibtex
