+++ paper/paper.bib
@@ -31,7 +31,7 @@
 
 @book{Wexler1983,
   title={Thermodynamic Properties of Dry Air, Moist Air and Water and SI Psychrometric Charts},
-  author={Wexler, A. and Hyland, R.W. and Stewart, R. and American Society of Heating, Refrigerating and Air-Conditioning Engineers},
+  author={Wexler, A. and Hyland, R.W. and Stewart, R. and {American Society of Heating, Refrigerating and Air-Conditioning Engineers}},
   year={1983},
   publisher={ASHRAE},
   address = {Atlanta, GA}
@@ -52,7 +52,7 @@
 
 @book{Ashrae2017_IP,
   title={ASHRAE Handbook Fundamentals 2017: Inch-Pound System Edition},
-  author={{American Society of Heating Refrigeration and Air Conditioning Engineers}},
+  author={{American Society of Heating, Refrigerating and Air-Conditioning Engineers}},
   year={2017},
   series={ASHRAE Handbook Fundamentals Inch-Pound System},
   publisher={ASHRAE},
@@ -62,7 +62,7 @@
 
 @book{Ashrae2017_SI,
   title={ASHRAE Handbook Fundamentals 2017: International System Edition},
-  author={{American Society of Heating Refrigeration and Air Conditioning Engineers}},
+  author={{American Society of Heating, Refrigerating and Air-Conditioning Engineers}},
   year={2017},
   series={ASHRAE Handbook Fundamentals International System},
   publisher={ASHRAE},
