### 2018-12-01 #1113 @dmey
Hi @kyleniemeyer, thank you for pointing this out. I have changed the format to PDF but I am having issues with its placement -- see my last proof. The image ends up on page 2 even though there is enough space to position it on page 1.

Also, a couple of other things that I have noticed that may be worth addressing in the documentation (https://joss.readthedocs.io/en/latest/submitting.html):

- Add a supported media-formats section.
- Explain how to resize figures (if at all possible)
- Explain how to force figure placement in text (if at all possible) a bit like in this case -- the `H` option in LaTeX.

Beside this, I have only received a notification of this issues when you mentioned me but none when `@whedon` open the issue. Is this the expected behaviour?

If you want, I can move any of the above to a separate issue if you think that's necessary/easier for you to handle. Thanks.
