### 2018-12-02 0835c29
remove tests against python dev versions form CI
