+++ .travis.yml
@@ -2,9 +2,7 @@ language: python
 dist: xenial
 python:
     - "3.6"
-    - "3.6-dev"
     - "3.7"
-    - "3.7-dev"
 
 before_install:
   - sudo apt-get install gfortran
