### 2018-12-02 819b8b2
update travis badge + limit deploy to only python 3.6
