+++ .travis.yml
@@ -5,7 +5,6 @@ python:
     - "3.7"
 
 before_install:
-  - echo $TRAVIS_PYTHON_VERSION
   - sudo apt-get install gfortran
   - pip install Sphinx sphinx_bootstrap_theme sphinx-autodoc-typehints cffi
 
@@ -27,5 +26,5 @@ deploy:
  github_token: $GITHUB_TOKEN
  local_dir: build/html/api-docs
  on:
-   all_branches: true
+   branch: master
    condition: $TRAVIS_PYTHON_VERSION == 3.6
