### 2018-12-02 d1454d7
Update CI badge + limit deployment of gh-pages to python 3.6 and master branch only. (#4)
