+++ .travis.yml
@@ -26,4 +26,5 @@ deploy:
  github_token: $GITHUB_TOKEN
  local_dir: build/html/api-docs
  on:
-   all_branches: true
+   branch: master
+   condition: $TRAVIS_PYTHON_VERSION == 3.6
