### 2018-12-02 11b3fdb
fix python snippet in example usage page
