### 2018-12-04 d5071e3
improve print output for example usage
