### 2018-12-04 568d29b
add ready made examples
