+++ examples/js/README.md
@@ -1,0 +1,30 @@
+## PsychroLib JavaScript example
+
+This is a simple ready-made example showing how to use PsychroLib in a simple program. The purpose of this simple program is to calculate and print to standard output the dew-point temperature from dry-bulb temperature and relative humidity using PsychroLib's `GetTDewPointFromRelHum` function.
+
+### Prerequisites
+
+- [Git](https://git-scm.com/)
+- [Node.js](https://nodejs.org/)
+
+### Supported platforms
+
+- Windows
+- macOS
+- Linux 
+- Unix
+
+### How to run this example
+
+From the command prompt, clone [PsychroLib](https://github.com/psychrometrics/psychrolib) and navigate to `examples/python`. Then run the example program:
+
+```
+node example.js
+```
+
+The output will be:
+
+```
+TDewPoint: 21.309397163661785 degree C
+```
+
