+++ examples/excel/README.md
@@ -1,0 +1,17 @@
+## PsychroLib Microsoft Excel example
+
+This is a simple ready-made example showing how to use PsychroLib in a simple program. The purpose of this simple program is to calculate and print to standard output the dew-point temperature from dry-bulb temperature and relative humidity using PsychroLib's `GetTDewPointFromRelHum` function.
+
+### Prerequisites
+
+- [Git](https://git-scm.com/)
+- [Microsoft Excel](https://products.office.com/excel)
+
+### Supported platforms
+
+- Windows
+- macOS
+
+### How to run this example
+
+TODO
