+++ examples/README.md
@@ -1,0 +1,9 @@
+## PsychroLib ready-made examples
+
+This is a set of simple ready-made example showing how to use PsychroLib in a simple program to calculate and print to standard output the dew-point temperature from dry-bulb temperature and relative humidity using PsychroLib's `GetTDewPointFromRelHum` function for all the languages we support.
+
+- [Python](python/README.md)
+- [C](c/README.md)
+- [Fortran](fortran/README.md)
+- [JavaScript](js/README.md)
+- [Excel](excel/README.md)
