### 2018-12-05 #1113 @kyleniemeyer
@bryanwweber do you have the bandwidth to review this? Looks similar to stuff you've worked on (and teach 😄 )
