### 2018-12-05 #1113 @dmey
Thank you both, @kyleniemeyer and @bryanwweber.
@kyleniemeyer, looking at the list of JOSS reviewers, `andreas-h`/`katyhuff`/`wholmgren` looked like they may be appropriate but if you had other colleagues in mind or you can ask that's really good too! Thanks for asking!
