### 2018-12-05 #1113 @dmey
Is everyone OK if I move the figure placement issue -- i.e. force the position of a figure so that it does not end up in between the reference section -- to https://github.com/openjournals/joss/issues?
