### 2018-12-05 #1113 @bryanwweber
@dmey Rather than using `H` to force the figure placement on the first page, you could consider moving the image up the page, to before the first paragraph and see what happens.
