### 2018-12-05 #1113 @dmey
@bryanwweber even when the figure is placed at the top of the page, the figure still ends up between the refs. I'll open an issue then.
