### 2018-12-05 7d58638
try moving fig to top of page
