### 2018-12-06 4263e05
set paper fig width to .9
