### 2018-12-06 #1113 @arfon
@dmey - the command to compile from a different branch doesn't seem to be working.
