### 2018-12-06 #1113 @dmey
@arfon I was just getting back to you in the issue about this. I think there are two issues here then. One with the `generate from` command to use for testing and the other with the actual figure placement.
