### 2018-12-06 #1113 @dmey
@arfon can you manually set the branch to `dmey/paper` so that we can use it to sort out these draft and during review? For the `generate from` command issue, I noticed that there is already openjournals/joss#471 open. Thanks.
