### 2018-12-06 #1113 @arfon
Sure, I can compile locally. Here's the current output of `dmey/paper`: [10.21105.joss.01113.pdf](https://github.com/openjournals/joss-reviews/files/2654212/10.21105.joss.01113.pdf)
