### 2018-12-06 #1113 @dmey
@arfon sorry but I meant if you can set `@whedon` to track `dmey/paper` as the default branch instead of `master` for now on so that we can carry out the review there and we would also not have to rely on the `generate from` command which is currently not working -- I would like to keep `master` tidy.
