### 2018-12-06 7cdc6e3
tentative resize fig [wip]
