### 2018-12-06 8e0ffc0
resize fig [wip]
