### 2018-12-06 f1adb26
resize fig [wip]
