### 2018-12-06 #1113 @dmey
@arfon the branch issue doesn't appear to be an issue after all -- silly me, I was using `generate pdf from <branch_name>` instead of `generate pdf from branch <branch_name>`... 😬. For reference though it may still be good to handle invalid commands in whedon.

I'd say that the latest formatting is now acceptable to review -- perhaps there may be other changes that can be addressed during the review process. If there are no issues with anyone, I would therefore keep the latest version of the paper under the `dmey/paper` branch.
