### 2018-12-08 dd3d9d6
Fixed docstring example for setting unit system
The current example generates the following error: 
`ValueError: The system of units has to be either SI or IP.`
