### 2018-12-08 73f46fa
Fortan: fix docsting in src file example
