+++ docs/sphinx/index.rst
@@ -1,0 +1,9 @@
+.. raw:: html
+   :file: ghbtns.html
+
+.. mdinclude:: ../../README.md
+.. toctree::
+   :hidden:
+   :maxdepth: 2
+
+   PsychroLib API<./api-docs>
