+++ docs/api-docs/index.rst
@@ -1,15 +1,0 @@
-=============================
- PsychroLib API Documentation
-=============================
-
-The page provides a comprehensive description of all the 
-functions available in PsychroLib. Although it describes 
-the Python version of the library, PsychroLib uses a 
-common API across all the supported language implementations.
-Please see the `PsychroLib repository on GitHub`_ if you are looking to
-download or find out more about PsychroLib.
-
-.. _`PsychroLib repository on GitHub`: https://github.com/psychrometrics/psychrolib
-
-.. automodule:: psychrolib
-   :members:
