### 2018-12-15 27af063
update doc config
