### 2018-12-20 #1137 @kyleniemeyer
👋  @dmey @bryanwweber @puncan

@bryanwweber @puncan please note your separate review checklists above; you can submit any other general comments as comments here, or as issues filed in the software repo (just make sure to either mention this issue in those, or comment with any issues you file here)
