### 2018-12-20 #1137 @dmey
Thank you @kyleniemeyer @bryanwweber @puncan.

Just to let you all know that we have had a few issues with the positioning of the image therefore I am currently using the [dmey/paper](https://github.com/psychrometrics/psychrolib/tree/dmey/paper) branch for this submission to avoid polluting master with multiple commits -- i.e. for generating the summary document or addressing any of the comments you may have. `@whedon` cannot set the branch to an arbitrary one therefore I have had to generate the summary manually -- for the current version of the summary please see [this](https://github.com/openjournals/joss-reviews/issues/1137#issuecomment-449064597).

One note with regards to the following point -- `Version: Does the release version given match the GitHub release (2.0.0)?` I have not made a release yet but will do so if/when this  has been finalised and added to Zenodo just in case there was anything else to add/modify.
