### 2019-01-03 8c680f4
fix VapPres bounds check
