### 2019-01-04 psychrometrics/psychrolib#8 @puncan
Test files do not run in Python
**Describe the bug**
When attempting to use PyTest to run the test files, each file independently reports:
ModuleNotFoundError: No module named 'psychrolib_fortran'

**To Reproduce**
Steps to reproduce the behavior:
1. Clone repo
2. Navigate to repo location in terminal
3. Enter: pytest test_psychrolib_si.py

**Expected behavior**
I expected a small report to be generated declaring how many tests had passed and failed.

 - OS: Ubuntu 18.04.01 LTS
 - Browser: Chrome
 - Version: 2.0.0

**Additional context**
I could be approaching the test files incorrectly. My experience is limited with PyTest.
