### 2019-01-04 #1137 @puncan
Other than the issue I just opened, I need the following checked off the list:

1. A statement of need in the general documentation, maybe paraphrase or pull from what is in the paper?
2. Authors and references in the paper.md file

I do see the bib file for the references but it should be used in the paper.md file.
