### 2019-01-04 psychrometrics/psychrolib#8 @dmey
Hi @puncan, the tests are automatically run at each commit directly on Continuous Integration (CI) -- see [here](https://travis-ci.com/psychrometrics/psychrolib/). For example if you want to see the tests results of the latest build you can see [these lines]( https://travis-ci.com/psychrometrics/psychrolib/jobs/167987957#L1258-L1394). If you however want to run these tests locally you will need to set up the environment correctly. I cannot see the full log in your issue but I presume you did not have installed the necessary dependencies to allow f2py to compile the Fortran library. For your specific system the following should work:

N.B. I assume that `pip` and `python` are aliases to pip-3.6 and python-3.6 version or greater

From your command prompt terminal
``` bash
sudo apt-get install gfortran
pip install m2r cffi numpy
# make sure your are at psychrolib's root dir
python -m pytest -v -s
```

If you have other issues please do let me know (and attach a full log please) but this should work on your local machine.
