### 2019-01-04 111e6ed
add development and testing notes.
