### 2019-01-04 #1137 @dmey
@puncan thanks for the comments. Just to let you know we are working on the [dmey/paper branch](https://github.com/psychrometrics/psychrolib/tree/dmey/paper) to avoid polluting master for now.

I have addressed the issue you opened (https://github.com/psychrometrics/psychrolib/issues/8) -- I have also just added a general readme for developing and testing purpose to mention how to run local testing in the `dmey/paper` branch (see [DEVELOP.md](https://github.com/psychrometrics/psychrolib/blob/dmey/paper/DEVELOP.md)). For the other points you raised, please see my responses below:
1. I have add this in the `README.md` file -- see [this](https://github.com/psychrometrics/psychrolib/tree/dmey/paper).
2. I am not sure about this. The `paper.md` [here](https://raw.githubusercontent.com/psychrometrics/psychrolib/dmey/paper/paper/paper.md) should contain all the information needed to generate the [paper above](https://github.com/openjournals/joss-reviews/issues/1137#issuecomment-449064597). Was this the one you were referring to?
