### 2019-01-06 #1137 @kyleniemeyer
@puncan regarding the paper contents, you can just review the generated PDF for the references and such.
