+++ src/fortran/psychrolib.f90
@@ -81,6 +81,8 @@ module psychrolib
   public :: GetStandardAtmTemperature
   public :: GetSeaLevelPressure
   public :: GetStationPressure
+  public :: GetSpecificHumFromHumRatio
+  public :: GetHumRatioFromSpecificHum
   public :: CalcPsychrometricsFromTWetBulb
   public :: CalcPsychrometricsFromTDewPoint
   public :: CalcPsychrometricsFromRelHum
@@ -1091,6 +1093,35 @@ module psychrolib
 
     StationPressure = SeaLevelPressure / GetSeaLevelPressure(1.0, Altitude, TDryBulb)
   end function GetStationPressure
+
+
+  !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+  ! Conversion between humidity types
+  !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+  function GetSpecificHumFromHumRatio(HumRatio) result(SpecificHum)
+    !+ Return the specific humidity from humidity ratio (aka mixing ratio).
+    !+ Reference: http://glossary.ametsoc.org/wiki/Specific_humidity
+
+    real, intent(in) :: HumRatio
+      !+ Humidity ratio in lb_H₂O lb_Dry_Air⁻¹ [IP] or kg_H₂O kg_Dry_Air⁻¹ [SI]
+    real             :: SpecificHum
+      !+ Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]
+
+    SpecificHum = HumRatio / (1.0 + HumRatio)
+  end function GetSpecificHumFromHumRatio
+
+  function GetHumRatioFromSpecificHum(SpecificHum) result(HumRatio)
+    !+ Return the humidity ratio (aka mixing ratio) from specific humidity.
+    !+ Reference: http://glossary.ametsoc.org/wiki/Specific_humidity
+
+    real, intent(in)  :: SpecificHum
+      !+ Humidity ratio in lb_H₂O lb_Air⁻¹ [IP] or kg_H₂O kg_Air⁻¹ [SI]
+    real              :: HumRatio
+      !+ Humidity ratio in lb_H₂O lb_Dry_Air⁻¹ [IP] or kg_H₂O kg_Dry_Air⁻¹ [SI]
+
+    HumRatio = SpecificHum / (1.0 - SpecificHum)
+  end function GetHumRatioFromSpecificHum
 
 
   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
