### 2019-01-07 #1137 @puncan
Okay, I see the PDF now, it looks great. Sorry about that! The readme also looks good, thank you for the update.
