### 2019-01-07 psychrometrics/psychrolib#8 @puncan
That worked very well, thank you for the added documentation and your comments.
