### 2019-01-08 #1137 @bryanwweber
Hi @dmey! I have a few comments. Please let me know if you want me to open issues for any of these:

1. Along with @puncan's comment, the documentation website doesn't provide a statement of need, and it would help if it did
2. For that matter, the documentation website is not linked from anywhere on the repository page that I could see
3. For me, the contributing instructions are not sufficient. It would help to have something more explicit, like the examples on the GitHub documentation: https://help.github.com/articles/setting-guidelines-for-repository-contributors/#examples-of-contribution-guidelines (specifically, the Rails or Open Government examples)
4. In the installation instructions, it would help to include a list of dependencies, or explicitly state that no other dependencies are needed, particularly for the C and Fortran interfaces.
5. The examples you include are fairly trivial, could you include some more complex ones demonstrating what your software is useful for?
6. It would help to note (especially on the examples) that the SI units are SI + degC, since the SI unit of temperature is Kelvin.
7. In the paper,
  ```markdown
  such as converting between dry-,
  ```
  could be
  ```latex
  such as converting between \mbox{dry-,}
  ```
  to prevent the line break after the hyphen

And I have some separate comments for discussion, but these are more "opinion-based" so you don't need to make any changes based on these comments, but I'm interested in your thoughts.

1. I'm curious why you chose to reimplement the equations in all of the relevant languages, rather than have one "reference" implementation (e.g., in C) and have wrappers for other languages. The latter seems to me the more typical path for scientific software, especially in Python.
2. For the Python module specifically, I'm curious why you chose to use the C-like syntax, rather than something more Pythonic. For instance, typically in Python, functions are named as all lower-case with underscores, or returning a dictionary rather than a rather large tuple.
3. Have you thought about including a separate units library to handle conversions, rather than having the global constants?
