### 2019-01-08 28fcd79
prevent the line break after the hyphen in paper
