### 2019-01-08 95d5de5
remove clutter / update readme and contributing instructions
