### 2019-01-08 8bd45ac
specifiy unit of temperature in examples
