### 2019-01-08 84de70a
remove bullet points from Copyright and license
