### 2019-01-08 1699d70
fix mistake in coding conventions section
