### 2019-01-08 #1137 @dmey
Hi @bryanwweber, many thanks for your time in reviewing PsychroLib and the useful comments. For clarity, I will include my responses ordered below.

*Assumption: you have checked-out the [dmey/paper](https://github.com/psychrometrics/psychrolib/tree/dmey/paper) branch.*

## Main comments

1. I think that this was addressed [here](https://github.com/psychrometrics/psychrolib/commit/f64f6bc6dfe8cd7680d622334b160a5d670208b0) -- the readme should now show that. Please can you let me know if this is OK with you or if  it requires further changes?

2. I have modified the readme and have linked to both the list of available functions and the API docs under [Documentation](https://github.com/psychrometrics/psychrolib/tree/dmey/paper#documentation) heading -- please let me know if this is still not clear and I will amend it.

3. I have modified this as you have suggested in my last commits. Please pull and let me know if you require more changes.

4. Done. See second paragraph under [Installing](https://github.com/psychrometrics/psychrolib/tree/dmey/paper#installing) in the readme.

5. I am not sure about this. PsychroLib is a general library for calculating psychometric properties of air therefore I think that the user should at least know a priori the type of functions to use in their study/software. I am not sure that more complex examples would substantially increase the user's understanding of what the library can/cannot do. For that I think that the list of functions and descriptions [here](https://github.com/psychrometrics/psychrolib/tree/dmey/paper/docs/available-functions.md) is more beneficial. Having said this, I had previously created some ready-made examples,please see [this](https://github.com/psychrometrics/psychrolib/blob/dmey/examples/examples/README.md) (as currently in another branch) that I could add if you think would be beneficial? But please let me know what you think. Thanks.

6. I have clarified it at the end of the first paragraph in the example usage page (see [this](https://github.com/psychrometrics/psychrolib/blob/dmey/paper/docs/how-to-use-psychrolib.md)).

7. Thanks. Done.

## Opinion-based comments

1. Calling C from Python and Fortran is somewhat trivial but not sure how easy it is to do it from JavaScript when running in browser and not sure it is even possible from VBA/Excel as far as I am aware. Having said this even in Python I would have had to create a boilerplate and had to rely on complied code making the distributions more challenging to create and maintain -- everything in PsychroLib is native for the specific language and there are no dependencies. For Python, for example, the installation via pip will 'just work' without having to rely on pre-compiled binaries for a specific platform or system. I think that the emphasis of PsychroLib at the moment is it's simplicity. Of course in some of my other larger projects (cf. [GIS4WRF](https://github.com/GIS4WRF/gis4wrf), [WRF-CMake](https://github.com/WRF-CMake)) I do things very differently however I think that in the case of PsychroLib it may be better to keep everything as simple as possible and keep supporting languages such as VBA/Excel where I find users relying upon... Hopefully this is a reasonable compromise to keep the implementations as simple as possible whilst also allowing users from different backgrounds to be able to approach, use, and integrate PsychroLib in their software.

2. Consistency. I agree and, must admit, I personally dislike when standards are not adhered to. In this case however, I think that using camel case across different languages (esp. in Python) was reasonable. We wanted to keep a consistent API across different languages therefore, I would argue that it makes it easier to compare, work, and understand functions with the same name across the different languages we support.

3. Good point but I think that this would only be trivial in Python (e.g. using `pint` or similar) but not in the other languages we support...

In general, I think that if we only wanted to support Python we could have chosen to write the libraries using language-specific features (e.g. some of the OO paradigms), but in PsychroLib we had to compromise in order to support various languages that users from different backgrounds could make use of.
