### 2019-01-08 #1137 @bryanwweber
Hi @dmey! Thanks for the updates. Some responses:

1. and 2. I was thinking to update the documentation page here: https://psychrometrics.github.io/psychrolib/index.html with the "Statement of Need" and also linking directly to that page somewhere on the repo page (recognizing the content is nearly the same as the README). I also think you should be consistent about linking to the rendered HTML pages and the rendered Markdown on GitHub (with my preference being for the former).

#3. Thanks, that looks good! One small thing in the test section, you should specify to install pytest as well.

#5. Those examples seem the same as the examples you've already got, except they're directly executable. I was thinking of an example that motivated you to write this software in the first place. What was a problem you were trying to solve that you used Psychrolib for? By the way, I'm interpreting this comment in light of the checkbox

> **Example usage**: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I agree that if you know you need this software, you probably know what you're doing, and this package has a simple enough API that it might not be necessary to have a very complicated example. But if you have something handy, I think it would improve the motivation to use this software.

And the other questions

1. Totally fair, I know very well that distributing C-libraries for Python is non-trivial, and it wouldn't help you with several of your other interfaces. Its good that you have a unified testing interface to make sure the interfaces don't fall out of sync.
2. Also fair. That said, do many people move from one language to the other and expect the same API? Just some food for thought for the future 😄
3. Yes, Python was what I had in mind here. It seems that JS has libraries (e.g. https://github.com/ben-ng/convert-units), but C, Fortran, and Excel would be more-or-less left out, unless you wrote it yourself. (There are many C++ units conversion libraries, but none I see that support just C.)

And one more comment that I edited in to my previous comment but must have been missed (sorry!)

4. In your tests, you tend to change the relative comparison threshold per-test, presumably you've figured out some small value that works for each test. I'm surprised to see errors as large as you say compared to the ASHRAE data, considering that you're implementing the ASHRAE formulae. Do you have any speculation why the error is so variable?
