### 2019-01-09 ed4149e
change ghpages deploy condition to dmey/paper
