### 2019-01-09 dcbffa5
use rel links only in docs
