### 2019-01-09 f030ebf
revert Travis to deploy gh-pages on master only
