### 2019-01-09 #1137 @dmey
Hi @bryanwweber, thanks. My responses:

## Main comments
1\. and 2\. Apologies but forgot to mention that the pages were generated automatically at commits but restricted to deploy only on `master` -- this is probably why you could not see the changes but @puncan could as he was probably looking directly at the `README.md`. In any case, I have simplified the process now and only generate the html docs and use gh-pages for the API docs (i.e. only for https://psychrometrics.github.io/psychrolib/api-docs.html) -- this will also also make it easier to maintain the docs in the future.

3\. Done. Thanks!

5\. I see. One of the main motivations for me was to use PsychroLib in physics-based/engineering models (e.g. energy balance models) I am working on/currently developing -- a common point of frustration was that these functions were not used/available consistently and, across the number of languages I commonly use (i.e. C, Fortran, and Python) or across models but others may have other uses such as to convert data (a very simple, but common example, for example is when the input to a model is in the form of humidity ratio instead of relative humidity as provided by weather stations).

----

4\. In the tests (e.g. `tests\test_psychrolib_si.py`) `test_GetStandardAtmPressure` and `test_GetStandardAtmTemperature` do have a high absolute error threshold -- 1 and 0.1 respectively -- but I think that this is reasonable given that in Table 1 of ch. 1 of the 2017 ASHRAE Handbook - Fundamentals the pressure is given in kPa -- the abs error of 1 refers to the last significant digit in the table, as it is for the temperature with 0.1. The temperature is only given in table 1 with 1 significant digit...
