### 2019-01-09 #1137 @dmey
@bryanwweber thanks again for your time and useful comments.

@kyleniemeyer when the changes are finalised, please can you let me know if I should merge the `dmey/paper` branch to `master` and create releases for the new version. Thanks
