### 2019-01-09 #1137 @kyleniemeyer
@dmey alright, looks good to me. Please merge the paper branch, create a new release, deposit in Zenodo for a DOI, and report the DOI here.
