### 2019-01-10 aeabd42
fix typos in CONTRIBUTING.md
