### 2019-01-11 b254b1b
JOSS review: update paper and readme  + add Zenodo metadata (#11)
