+++ docs/sphinx/conf.py
@@ -23,9 +23,9 @@ import psychrolib
 
 # -- Project information -----------------------------------------------------
 
-project = 'PsychroLib'
-copyright = '2018, D. Thevenard and D. Meyer'
-author = 'D. Thevenard and D. Meyer'
+project = u'PsychroLib'
+copyright = u'2018, D. Thevenard and D. Meyer'
+author = u'D. Thevenard and D. Meyer'
 
 # -- General configuration ---------------------------------------------------
 
@@ -44,7 +44,7 @@ extensions = ['sphinx.ext.autodoc',
               'm2r']
 
 # Add any paths that contain templates here, relative to this directory.
-#templates_path = ['_templates']
+templates_path = ['_templates']
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
@@ -52,7 +52,7 @@ extensions = ['sphinx.ext.autodoc',
 source_suffix = ['.rst', '.md']
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = 'api-docs'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
