+++ README.md
@@ -1,98 +1,92 @@
-<p align="center"><img src="https://raw.githubusercontent.com/psychrometrics/psychrolib/master/assets/psychrolib-logo.svg?sanitize=true" alt="PsychroLib Logo" height="150" width="150"></p>
+<p align="center"><img src="assets/psychrolib-logo.svg" alt="PsychroLib Logo" height="150" width="150"></p>
 
-# PsychroLib - A Library of Psychrometric Functions in C, Fortran, VBA (Excel), JavaScript, and Python
-[![Build Status](https://travis-ci.com/psychrometrics/psychrolib.svg?branch=master)](https://travis-ci.com/psychrometrics/psychrolib)
+# PsychroLib: a library of psychrometric functions for Python, C, Fortran, JavaScript and VBA/Excel [![Build Status](https://travis-ci.com/psychrometrics/psychrolib.svg?branch=master)](https://travis-ci.com/psychrometrics/psychrolib)
+
+
+### Table of contents
+- [Introduction](#introduction)
+- [Documentation](#documentation)
+- [Installing](#installing)
+- [Example usage](#example-usage)
+- [Contributing](#contributing)
+- [Versioning](#versioning)
+- [Copyright and license](#copyright-and-license)
+- [Acknowledgements](#acknowledgements)
+
+
+## Introduction
 
 Psychrometrics are the study of physical and thermodynamic properties of moist air. These properties include, for example, the air's dew point temperature, its wet bulb temperature, relative humidity, humidity ratio, enthalpy.
 
+The estimation of these properties is critical in several engineering and scientific applications such as heating, ventilation, and air conditioning (HVAC) and meteorology. Although formulae to calculate the psychrometric properties of air are widely available in the literature ([@Stull2011]; [@Wexler1983]; [@Stoecker1982]; [@Dilley1968]; [@Humphreys1920]), their implementation in computer programs or spreadsheets can be challenging and time consuming.
+
 PsychroLib is a library of functions to enable calculating psychrometric properties of moist and dry air. The library is available for Python, C, Fortran, JavaScript, and Microsoft Excel Visual Basic for Applications (VBA). The functions are based of formulae from the  2017 ASHRAE Handbook — Fundamentals, Chapter 1. Functions can be grouped into two categories:
 
-1. Functions for the calculation of dew point temperature,
-    wet-bulb temperature, partial vapour pressure of water, humidity
-    ratio or relative humidity, knowing any other of these and dry bulb
-    temperature and atmospheric pressure;
+1. Functions for the calculation of dew point temperature, wet-bulb temperature, partial vapour pressure of water, humidity ratio or relative humidity, knowing any other of these and dry bulb temperature and atmospheric pressure.
 
-2. Functions for the calculation of other moist air properties. All these use the
-    humidity ratio as input.
+2. Functions for the calculation of other moist air properties. All these use the humidity ratio as input.
 
[…]

[…]
