### 2019-01-11 #1137 @dmey
Thanks @puncan and @kyleniemeyer

@kyleniemeyer I have two DOIs:
- One that represents all versions, and will always resolve to the latest one: [10.5281/zenodo.2537945](https://dx.doi.org/10.5281/zenodo.2537945)
- One that represents this specific version -- in this case is the latest and only one: [10.5281/zenodo.2537946](https://dx.doi.org/10.5281/zenodo.2537946)

I looked at the JOSS docs (specifically [here](https://joss.readthedocs.io/en/latest/editing.html?highlight=doi#after-acceptance)) but could not anything specific about this -- I think it would also be useful to mention this in the docs -- if you want I can also open an issue for this. Thanks
