### 2019-01-11 #1137 @kyleniemeyer
@dmey For this case, we want the paper to reference the specific version of the software that was reviewed, so I'll use that DOI.
