### 2019-01-13 d199f2d
add citing section and badges
